+++ doc/changelog.md
@@ -3,14 +3,15 @@
 
 ## Work In Progress
 
-- Renamed Puffs to Wuffs
+- Renamed Puffs to Wuffs.
+- Added a `yield` keyword.
 
 
 ## 2017-11-16
 
 - [Initial open source
   release](https://groups.google.com/d/topic/puffslang/2z61mNTAMns/discussion),
-  under the "Puffs" name
+  under the "Puffs" name.
 
 
 ---